Update default server ports
The configuration for default server ports in the application properties was modified. The quarkus.http.port in the paw-patrol project was set to 8082 to avoid port conflicts when running multiple services locally. Similarly, the server.port in the main resources was changed to 8081. Both updates help to streamline local development.