Add Paw entity and PawResource endpoint
Introduced a new Paw entity along with a 'GET' resource endpoint in PawResource. The Paw entity includes fields for name, description, and fun fact which can be accessed via the endpoint. This addition allows retrieval of a list of all Paw entities stored in the database, improving the overall functionality of the application.