Add data import script for 'paw' table
Introduced 'import.sql' that includes data for the 'paw' table. This SQL file deletes all entries from the 'paw' table and subsequently inserts a set of predefined entries. This will be particularly useful in development and testing environments, allowing for consistent and relevant data to be imported and worked with.