Add a random Paw feature to API endpoints
In this commit, a random Paw retrieval functionality is added to both the Paw model and PawResource. A method findRandom() utilizing java.util.Random is introduced in the Paw.java - for random Paw selection. A corresponding API endpoint '/random' is added in PawResource.java to expose this functionality via API call, thereby giving opportunity for the clients to fetch random Paw data.