Added PostgreSQL datasource details and uber-jar packaging
In this commit, PostgreSQL has been set as the default database kind along with     username and password details. Additionally, an Uberjar packaging method is defined for the Quarkus application. The change was necessary to configure PostgreSQL as the project database and pack the application into a standalone, executable JAR file with all its dependencies.

Also, the configuration of the Spring application in the dinosaur module has been changed to include PostgreSQL connection details. Furthermore, virtual threads are enabled for Spring concurrency model optimization.